pyproject.toml
[tool.poetry]
name = "ctxlib3"
version = "0.1.0"
description = ""
authors = ["Philip Solarz <philip.solarz@teliacompany.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.4.2"
docarray = {extras = ["redis"], version = "^0.39.0"}
fastapi = "^0.104.0"
llama-cpp-python = "^0.2.11"
langchain = "^0.0.317"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
